rearrange assets and main.js to make them run with vite
