+++ src/platform/web/assets/config.json
