+++ src/platform/web/assets/download-sandbox.html
