+++ src/platform/web/assets/icon-maskable.png
