+++ src/platform/web/assets/icon.png
