+++ src/platform/web/assets/manifest.json
