+++ vite.config.js
@@ -1,0 +1,7 @@
+export default {
+    public: false,
+    root: "src/platform/web",
+    server: {
+        hmr: false
+    }
+};
